fix: move folders outside app into src
