+++ src/app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
-import { ApolloWrapper } from '@/app/lib/ApolloClient';
-import { ProductProvider } from '@/app/context/ProductContext';
-import { CartProvider } from '@/app/context/CartContext';
+import { ApolloWrapper } from '@/lib/ApolloClient';
+import { ProductProvider } from '@/context/ProductContext';
+import { CartProvider } from '@/context/CartContext';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
