+++ src/components/Product/ProductCard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Product } from '@/app/graphql/types';
+import { Product } from '@/graphql/types';
 import { useMutation } from '@apollo/client';
-import { ADD_ITEM_MUTATION } from '@/app/graphql/mutations';
+import { ADD_ITEM_MUTATION } from '@/graphql/mutations';
 
 interface ProductProps {
   product: Product;
